Go back to OpenGL 3.3 and crash correctly on Windows
On Windows, an error box is now shown when the game crashes at initialization (TODO: wrap main() so that a box is also shown for any error thrown)

The game has been retrograded to OpenGL 3.3 which allows it to run on much more platforms while (currently) preventing no features. Each OpenGL version bump will now have to be working on most playtesters' computers.