fix: Prevent water from disappearing
The bug was caused by the fact that `sendWater` could send water to a point that's loaded in the current vector. The problem is that the current vector will overwrite newWaterMass. So in the end the transmitted water was nulled.

This caused water to disappear but has now been fixed (albeit the function is slower now)